examples/order/python/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1,<0.112.0
uvicorn[standard]>=0.24.0,<0.31.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6,<0.1.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0

# Logging
structlog>=23.2.0,<25.0.0

# OpenTelemetry
opentelemetry-api>=1.21.0,<2.0.0
opentelemetry-sdk>=1.21.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.42b0,<1.0.0
opentelemetry-instrumentation-logging>=0.42b0,<1.0.0
opentelemetry-exporter-jaeger-thrift>=1.21.0,<2.0.0

# Events
cloudevents>=1.10.1,<2.0.0

# Configuration
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0,<7.0.0

# Testing
pytest>=7.4.3,<9.0.0
pytest-asyncio>=0.21.1,<1.0.0
httpx>=0.25.2,<1.0.0
